simple calculator challenge ch1

+++ simplecalculater/simplecalculater/Default.aspx.cs
@@ -51,9 +51,9 @@ namespace simplecalculater
 
         protected void divButton_Click(object sender, EventArgs e)
         {
-            double value1 = Convert.ToDouble(num1TextBox.Text);
-            double value2 = Convert.ToDouble(num2TextBox.Text);
-            double result = value1 / value2;
+            decimal value1 = Convert.ToDecimal(num1TextBox.Text);
+            decimal value2 = Convert.ToDecimal(num2TextBox.Text);
+            decimal result = value1 / value2;
 
             resultLabel.Text = result.ToString();
         }
